Initial modules
Added some of the initial modules. The code does not work yet, but
the modules are creaetd. The fsutils.ts is solid, but needs tests.

The todo.ts and extension.ts still need a good design session, drawing
the goal of the module.

Started adding configuration to the contribution in the package.json,
however, this also still needs to be designed properly.